+++ .eslintrc.json
@@ -11,7 +11,6 @@
   },
   "extends": ["airbnb-base"],
   "rules": {
-    "max-classes-per-file":0,
     "no-shadow": "off",
     "no-param-reassign": "off",
     "eol-last": "off",
