+++ module/constructor.js
@@ -1,0 +1,7 @@
+/* eslint-disable linebreak-style */
+export default class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
